added continue button to main screen added proxy ai and empty lobby menu reworked some network stuff added network settings to the option menu more battle sprites added

+++ menue/victorymenue.cpp
@@ -147,11 +147,11 @@ VictoryMenue::VictoryMenue()
 
     if (pMap->getCurrentDay() > 1)
     {
-        lineLength = static_cast<qint32>(m_pGraphBackground->getWidth() / (pMap->getCurrentDay() - 1));
+        lineLength = m_pGraphBackground->getWidth() / static_cast<float>(pMap->getCurrentDay() - 1);
     }
     else
     {
-        lineLength = static_cast<qint32>(m_pGraphBackground->getWidth());
+        lineLength = m_pGraphBackground->getWidth();
     }
 
     m_ProgressTimer.setSingleShot(false);
